Made everything compatible with Fragments
