+++ app/src/main/java/com/example/roomieprototype/CardStack.java
@@ -8,6 +8,7 @@ import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.BaseAdapter;
+import android.widget.FrameLayout;
 import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
@@ -38,63 +39,32 @@ import java.util.concurrent.TimeoutException;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.fragment.app.Fragment;
 
 import static com.google.android.gms.tasks.Tasks.await;
 
-public class CardStack extends AppCompatActivity implements SwipeStack.SwipeStackListener, View.OnClickListener {
 
-    public String imgStr;
-
-    private ArrayList<String> mData;
-    private SwipeStack mSwipeStack;
-    private SwipeStackAdapter mAdapter;
-    public int count;
-    private FloatingActionButton mButtonLeft, mButtonRight, mRewind;
-    private DrawerLayout drawerLayout;
-    private Query firebaseUsers;
-    private FirebaseFirestore db;
-    private FirebaseUser user;
-    private String userSleep;
-    private String userClean;
-    private String userEat;
-    private String userStudy;
-    private String userSocial;
-    private String userTemperature;
-    private String userApart;
-    private String userDorm;
-    private ArrayList<String> matchList;
-  
 public class CardStack extends AppCompatActivity implements View.OnClickListener, BottomNavigationView.OnNavigationItemSelectedListener {
 
     public BottomAppBar bottomAppBar;
     BottomNavigationView bottomNavigationView;
 
     private FloatingActionButton mBottomFAB;
+    private ArrayList<String> matchList;
+    private FragmentMatch fragobj;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.main_activity);
+
         matchList = getIntent().getStringArrayListExtra("matchList");
-        count = 0;
-
-        mSwipeStack = findViewById(R.id.swipeStack);
-        mButtonLeft = findViewById(R.id.skip_button);
-        mButtonRight = findViewById(R.id.like_button);
-        mRewind = findViewById(R.id.rewind_button);
-
-        mButtonLeft.setOnClickListener(this);
-        mButtonRight.setOnClickListener(this);
-        mRewind.setOnClickListener(this);
-
-        mData = new ArrayList<>();
-        mAdapter = new SwipeStackAdapter(mData,matchList);
-        mSwipeStack.setAdapter(mAdapter);
-        mSwipeStack.setListener(this);
-
-        //mClose = findViewById(R.id.close_button);
-
-        fillWithTestData();
+        Log.d("TAG","cardstack"+matchList.toString());
+        Bundle bundle = new Bundle();
+        bundle.putStringArrayList("matchList", matchList);
+        // set Fragmentclass Arguments
+        fragobj = new FragmentMatch();
+        fragobj.setArguments(bundle);
 
         bottomNavigationView = findViewById(R.id.bottom_navigation);
         bottomNavigationView.setOnNavigationItemSelectedListener(this);
@@ -103,7 +73,7 @@ public class CardStack extends AppCompatActivity implements View.OnClickListener
         // Set initial fragment
         if(savedInstanceState == null) {
             getSupportFragmentManager()
-                    .beginTransaction().setCustomAnimations(R.anim.fade_in, R.anim.fade_out).replace(R.id.container, fragmentMatch).addToBackStack(null).commitAllowingStateLoss();
+                    .beginTransaction().setCustomAnimations(R.anim.fade_in, R.anim.fade_out).replace(R.id.container, fragobj).addToBackStack(null).commitAllowingStateLoss();
         }
 
         mBottomFAB = findViewById(R.id.appbar_fab);
@@ -122,22 +92,7 @@ public class CardStack extends AppCompatActivity implements View.OnClickListener
 
     }
 
-    public void fillWithTestData() {
-        for (int x = 0; x < matchList.size(); x++) {
-            imgStr = "drawable/pic" + (x + 1);
-            mData.add(imgStr);
-        }
-    }
 
-    public String imageSwitch() {
-        count++;
-
-        if (count == matchList.size()+1) {
-            count = 1;
-        }
-        String uri = "drawable/pic" + count;
-        return uri;
-    }
 
     @Override
     public void onClick(View v) {
@@ -176,67 +131,7 @@ public class CardStack extends AppCompatActivity implements View.OnClickListener
         return false;
     }
 
-    @Override
-    public void onViewSwipedToRight(int position) {
-        String swipedElement = mAdapter.getItem(position);
-        Toast.makeText(this, "Swiped right",
-                Toast.LENGTH_SHORT).show();
-    }
 
-    @Override
-    public void onViewSwipedToLeft(int position) {
-        String swipedElement = mAdapter.getItem(position);
-        Toast.makeText(this, "Swiped left",
-                Toast.LENGTH_SHORT).show();
-    }
 
-    @Override
-    public void onStackEmpty() {
-        Toast.makeText(this, R.string.stack_empty, Toast.LENGTH_SHORT).show();
-    }
 
-    public class SwipeStackAdapter extends BaseAdapter {
-
-        private List<String> mData;
-        private List<String> mText;
-
-        public SwipeStackAdapter(List<String> data1,List<String> data2 ) {
-            this.mData = data1;
-            this.mText = data2;
-        }
-
-        @Override
-        public int getCount() {
-            return mData.size();
-        }
-
-        @Override
-        public String getItem(int position) {
-            return mData.get(position);
-        }
-
-        @Override
-        public long getItemId(int position) {
-            return position;
-        }
-
-        @Override
-        public View getView(final int position, View convertView, ViewGroup parent) {
-            if (convertView == null) {
-                convertView = getLayoutInflater().inflate(R.layout.swipe_card, parent, false);
-            }
-
-            ImageView imgViewCard = convertView.findViewById(R.id.imgView);
-
-            int imageResource = getResources().getIdentifier(mData.get(position), null, getPackageName());
-            Drawable image = getResources().getDrawable(imageResource);
-
-            TextView textViewCard = (TextView) convertView.findViewById(R.id.textViewCard);
-            textViewCard.setText(matchList.get(position));
-
-            imgViewCard.setImageDrawable(image);
-
-            return convertView;
-        }
-    }
 }
